Separate USBClient driver and STM32F4
