+++ Targets/STM32F4xx/BuildConfiguration.txt
@@ -1,2 +1,2 @@
 TargetArchitecture:CortexM4
-AdditionalTargetDrivers:SPIDisplay
+AdditionalTargetDrivers:SPIDisplay,USBClient
