+++ Targets/STM32F4xx/STM32F4_Startup.cpp
@@ -46,7 +46,7 @@ void STM32F4_Startup_OnSoftReset(const TinyCLR_Api_Provider* apiProvider) {
     STM32F4_Uart_Reset();
 #endif
 #ifdef INCLUDE_USBCLIENT 
-    STM32F4_UsbClient_Reset();
+    UsbClient_Reset();
 #endif
 }
 
@@ -451,7 +451,7 @@ void STM32F4_Startup_GetDebuggerTransportProvider(const TinyCLR_Api_Info*& api, 
     controller->ReleasePin(controller, DEBUGGER_SELECTOR_PIN);
 
     if (value == DEBUGGER_SELECTOR_USB_STATE) {
-        api = STM32F4_UsbClient_GetApi();
+        api = UsbClient_GetApi();
         index = USB_DEBUGGER_INDEX;
     }
     else {
