+++ Targets/STM32F4xx/STM32F4_USBC.cpp
@@ -16,1668 +16,8 @@
 
 #include <string.h>
 #include "STM32F4.h"
+#include "../../Drivers/USBClient/USBClient.h"
 
-#define __min(a,b)  (((a) < (b)) ? (a) : (b))
-
-#if defined(__GNUC__)
-#define PACKED(x) x __attribute__((packed))
-#elif defined(arm) || defined(__arm)
-#define PACKED(x) __packed x
-#endif
-
-///////////////////////////////////////////////////////////////////////////////////////////
-/// USB Debugger driver
-///////////////////////////////////////////////////////////////////////////////////////////
-
-// USB 2.0 host requests
-#define USB_GET_STATUS           0
-#define USB_CLEAR_FEATURE        1
-#define USB_SET_FEATURE          3
-#define USB_SET_ADDRESS          5
-#define USB_GET_DESCRIPTOR       6
-#define USB_SET_DESCRIPTOR       7
-#define USB_GET_CONFIGURATION    8
-#define USB_SET_CONFIGURATION    9
-
-// USB 2.0 defined descriptor types
-#define USB_DEVICE_DESCRIPTOR_TYPE        1
-#define USB_CONFIGURATION_DESCRIPTOR_TYPE 2
-#define USB_STRING_DESCRIPTOR_TYPE        3
-#define USB_INTERFACE_DESCRIPTOR_TYPE     4
-#define USB_ENDPOINT_DESCRIPTOR_TYPE      5
-
-// USB 2.0 host request type defines
-#define USB_SETUP_RECIPIENT(n)          ((n) & 0x0F)
-#define USB_SETUP_RECIPIENT_DEVICE             0x00
-#define USB_SETUP_RECIPIENT_INTERFACE          0x01
-#define USB_SETUP_RECIPIENT_ENDPOINT           0x02
-
-// Local device status defines
-#define USB_STATUS_DEVICE_SELF_POWERED   0x0001
-#define USB_STATUS_DEVICE_REMOTE_WAKEUP  0x0002
-
-#define USB_STATUS_ENDPOINT_HALT         0x0001
-
-#define USB_FEATURE_ENDPOINT_HALT        0x0000
-#define USB_FEATURE_DEVICE_REMOTE_WAKEUP 0x0001
-
-// Local device possible states
-#define USB_DEVICE_STATE_DETACHED       0
-#define USB_DEVICE_STATE_ATTACHED       1
-#define USB_DEVICE_STATE_POWERED        2
-#define USB_DEVICE_STATE_DEFAULT        3
-#define USB_DEVICE_STATE_ADDRESS        4
-#define USB_DEVICE_STATE_CONFIGURED     5
-#define USB_DEVICE_STATE_SUSPENDED      6
-#define USB_DEVICE_STATE_UNINITIALIZED  0xFF
-
-// Possible responses to host requests
-#define USB_STATE_DATA                  0
-#define USB_STATE_STALL                 1
-#define USB_STATE_DONE                  2
-#define USB_STATE_ADDRESS               3
-#define USB_STATE_STATUS                4
-#define USB_STATE_CONFIGURATION         5
-#define USB_STATE_REMOTE_WAKEUP         6
-
-
-// ATTENTION:
-// 2.0 is the lowest version that works with WinUSB on Windows 8!!!
-// use older values below if you do not care about that
-
-#define DEVICE_RELEASE_VERSION              0x0200
-
-//string descriptor
-#define USB_STRING_DESCRIPTOR_SIZE          32
-
-// index for the strings
-#define MANUFACTURER_NAME_INDEX             1
-#define PRODUCT_NAME_INDEX                  2
-#define SERIAL_NUMBER_INDEX                 0
-
-// configuration for extended descriptor
-#define OS_DESCRIPTOR_EX_VERSION            0x0100
-
-#define USB_DISPLAY_STRING_NUM     4
-#define USB_FRIENDLY_STRING_NUM    5
-
-#define OS_DESCRIPTOR_STRING_INDEX        0xEE
-#define OS_DESCRIPTOR_STRING_VENDOR_CODE  0xA5
-
-// USB 2.0 response structure lengths
-#define USB_DEVICE_DESCRIPTOR_LENGTH             18
-#define USB_CONFIGURATION_DESCRIPTOR_LENGTH       9
-#define USB_STRING_DESCRIPTOR_HEADER_LENGTH       2
-
-
-// USB configuration list structures
-#define USB_END_DESCRIPTOR_MARKER           0x00
-#define USB_DEVICE_DESCRIPTOR_MARKER        0x01
-#define USB_CONFIGURATION_DESCRIPTOR_MARKER 0x02
-#define USB_STRING_DESCRIPTOR_MARKER        0x03
-#define USB_GENERIC_DESCRIPTOR_MARKER       0xFF
-
-// configuration Descriptor
-#define USB_ATTRIBUTE_REMOTE_WAKEUP    0x20
-#define USB_ATTRIBUTE_SELF_POWER       0x40
-#define USB_ATTRIBUTE_BASE             0x80
-
-// Endpoint Direction
-#define USB_ENDPOINT_DIRECTION_IN 0x80
-#define USB_ENDPOINT_DIRECTION_OUT 0x00
-#define USB_ENDPOINT_NULL 0xFF
-
-// Endpoint Attribute
-#define ENDPOINT_INUSED_MASK        0x01
-#define ENDPOINT_DIR_IN_MASK        0x02
-#define ENDPOINT_DIR_OUT_MASK       0x04
-
-#define USB_ENDPOINT_ATTRIBUTE_BULK 2
-#define USB_MAX_DATA_PACKET_SIZE 64
-
-#define USB_MAX_EP_SIZE            64 // maximum control channel packet size
-#define USB_MAX_EP0_SIZE            8 // default control channel packet size
-#define USB_FS_MAX_EP_COUNT         4 // OTG FS supports 4 endpoints
-
-// This version of the USB code supports only one language - which
-// is not specified by USB configuration records - it is defined here.
-// This is the String 0 descriptor.This array includes the String descriptor
-// header and exactly one language.
-
-#define USB_LANGUAGE_DESCRIPTOR_SIZE 4
-
-// USB 2.0 request packet from host
-PACKED(struct) USB_SETUP_PACKET {
-    uint8_t bmRequestType;
-    uint8_t bRequest;
-    uint16_t wValue;
-    uint16_t wIndex;
-    uint16_t wLength;
-};
-
-PACKED(struct) USB_DYNAMIC_CONFIGURATION;
-
-struct USB_PACKET64 {
-    uint32_t Size;
-    uint8_t  Buffer[USB_MAX_DATA_PACKET_SIZE];
-};
-
-struct USB_PIPE_MAP {
-    uint8_t RxEP;
-    uint8_t TxEP;
-};
-
-struct USB_CONTROLLER_STATE;
-
-typedef void(*USB_NEXT_CALLBACK)(USB_CONTROLLER_STATE*);
-
-struct USB_CONTROLLER_STATE {
-    bool                                                        initialized;
-    uint8_t                                                     currentState;
-    uint8_t                                                     controllerNum;
-    uint32_t                                                    event;
-
-    const USB_DYNAMIC_CONFIGURATION*                            configuration;
-
-    /* queues & maxPacketSize must be initialized by the HAL */
-    USB_PACKET64                                   	            *queues[STM32F4_USB_QUEUE_SIZE];
-    uint8_t                                                     currentPacketOffset[STM32F4_USB_QUEUE_SIZE];
-    uint8_t                                                     maxPacketSize[STM32F4_USB_QUEUE_SIZE];
-    bool                                                        isTxQueue[STM32F4_USB_QUEUE_SIZE];
-
-    /* Arbitrarily as many pipes as endpoints since that is the maximum number of pipes
-       necessary to represent the maximum number of endpoints */
-    USB_PIPE_MAP                                                pipes[STM32F4_USB_QUEUE_SIZE];
-
-    /* used for transferring packets between upper & lower */
-    uint8_t*                                                    ptrData;
-    uint8_t                                                     dataSize;
-
-    /* USB hardware information */
-    uint8_t                                                     address;
-    uint8_t                                                     deviceState;
-    uint8_t                                                     packetSize;
-    uint8_t                                                     configurationNum;
-    uint32_t                                                    firstGetDescriptor;
-
-    /* USB status information, used in
-       GET_STATUS, SET_FEATURE, CLEAR_FEATURE */
-    uint16_t                                                    deviceStatus;
-
-    uint16_t                                                    endpointType;
-    uint16_t*                                                   endpointStatus;
-    uint8_t                                                     endpointCount;
-    uint8_t                                                     endpointStatusChange;
-
-    /* callback function for getting next packet */
-    USB_NEXT_CALLBACK                                           dataCallback;
-
-    /* for helping out upper layer during callbacks */
-    uint8_t*                                                    residualData;
-    uint16_t                                                    residualCount;
-    uint16_t                                                    expected;
-};
-
-PACKED(struct) TinyCLR_UsbClient_DescriptorHeader {
-    uint8_t  marker;
-    uint8_t  iValue;
-    uint16_t size;
-};
-
-PACKED(struct) TinyCLR_UsbClient_GenericDescriptorHeader {
-    TinyCLR_UsbClient_DescriptorHeader header;
-
-    uint8_t  bmRequestType;
-    uint8_t  bRequest;
-    uint16_t wValue;
-    uint16_t wIndex;
-};
-
-PACKED(struct) TinyCLR_UsbClient_DeviceDescriptor {
-    TinyCLR_UsbClient_DescriptorHeader header;
-
-    uint8_t  bLength;
-    uint8_t  bDescriptorType;
-    uint16_t bcdUSB;
-    uint8_t  bDeviceClass;
-    uint8_t  bDeviceSubClass;
-    uint8_t  bDeviceProtocol;
-    uint8_t  bMaxPacketSize0;
-    uint16_t idVendor;
-    uint16_t idProduct;
-    uint16_t bcdDevice;
-    uint8_t  iManufacturer;
-    uint8_t  iProduct;
-    uint8_t  iSerialNumber;
-    uint8_t  bNumConfigurations;
-};
-
-PACKED(struct) TinyCLR_UsbClient_InterfaceDescriptor {
-    uint8_t bLength;
-    uint8_t bDescriptorType;
-    uint8_t bInterfaceNumber;
-    uint8_t bAlternateSetting;
-    uint8_t bNumEndpoints;
-    uint8_t bInterfaceClass;
-    uint8_t bInterfaceSubClass;
-    uint8_t bInterfaceProtocol;
-    uint8_t iInterface;
-};
-
-PACKED(struct) TinyCLR_UsbClient_EndpointDescriptor {
-    uint8_t  bLength;
-    uint8_t  bDescriptorType;
-    uint8_t  bEndpointAddress;
-    uint8_t  bmAttributes;
-    uint16_t wMaxPacketSize;
-    uint8_t  bInterval;
-};
-
-PACKED(struct) TinyCLR_UsbClient_ClassDescriptorHeader {
-    uint8_t bLength;
-    uint8_t bDescriptorType;
-};
-
-PACKED(struct) TinyCLR_UsbClient_StringDescriptorHeader {
-    TinyCLR_UsbClient_DescriptorHeader header;
-
-    uint8_t bLength;
-    uint8_t bDescriptorType;
-    wchar_t stringDescriptor[32];
-};
-
-PACKED(struct) TinyCLR_UsbClient_ConfigurationDescriptor {
-    TinyCLR_UsbClient_DescriptorHeader header;
-
-    uint8_t  bLength;
-    uint8_t  bDescriptorType;
-    uint16_t wTotalLength;
-    uint8_t  bNumInterfaces;
-    uint8_t  bConfigurationValue;
-    uint8_t  iConfiguration;
-    uint8_t  bmAttributes;
-    uint8_t  bMaxPower;
-
-    TinyCLR_UsbClient_InterfaceDescriptor   itfc0;
-    TinyCLR_UsbClient_EndpointDescriptor    epWrite;
-    TinyCLR_UsbClient_EndpointDescriptor    epRead;
-};
-
-PACKED(struct) TinyCLR_UsbClient_OsStringDescriptor {
-    TinyCLR_UsbClient_DescriptorHeader header;
-
-    uint8_t   bLength;
-    uint8_t   bDescriptorType;
-    wchar_t signature[7];
-    uint8_t   bMS_VendorCode;
-    uint8_t   padding;
-};
-
-PACKED(struct) TinyCLR_UsbClient_XCompatibleOsId {
-    TinyCLR_UsbClient_GenericDescriptorHeader header;
-
-    uint32_t dwLength;
-    uint16_t bcdVersion;
-    uint16_t wIndex;
-    uint8_t  bCount;
-    uint8_t  padding1[7];
-    uint8_t  bFirstInterfaceNumber;
-    uint8_t  reserved;
-    uint8_t  compatibleID[8];
-    uint8_t  subCompatibleID[8];
-    uint8_t  padding2[6];
-};
-
-PACKED(struct) TinyCLR_UsbClient_XPropertiesOsWinUsb {
-    TinyCLR_UsbClient_GenericDescriptorHeader header;
-
-    uint32_t dwLength;
-    uint16_t bcdVersion;
-    uint16_t wIndex;
-    uint16_t  bCount;
-
-    uint32_t dwSize;
-    uint32_t dwPropertyDataType;
-    uint16_t wPropertyNameLengh;
-    uint8_t  bPropertyName[40];
-    uint32_t dwPropertyDataLengh;
-    uint8_t  bPropertyData[78];
-};
-
-PACKED(struct) USB_DYNAMIC_CONFIGURATION {
-    TinyCLR_UsbClient_DeviceDescriptor                  *device;
-    TinyCLR_UsbClient_ConfigurationDescriptor           *config;
-    TinyCLR_UsbClient_StringDescriptorHeader            *manHeader;
-    TinyCLR_UsbClient_StringDescriptorHeader            *prodHeader;
-    TinyCLR_UsbClient_StringDescriptorHeader            *displayStringHeader;
-    TinyCLR_UsbClient_StringDescriptorHeader            *friendlyStringHeader;
-    TinyCLR_UsbClient_OsStringDescriptor                *OS_String;
-    TinyCLR_UsbClient_XCompatibleOsId                   *OS_XCompatible_ID;
-    TinyCLR_UsbClient_XPropertiesOsWinUsb               *OS_XProperty;
-    TinyCLR_UsbClient_DescriptorHeader                  *endList;
-};
-
-USB_DYNAMIC_CONFIGURATION UsbDefaultConfiguration;
-
-const TinyCLR_UsbClient_DescriptorHeader * USB_FindRecord(USB_CONTROLLER_STATE* usbState, uint8_t marker, USB_SETUP_PACKET * iValue);
-void STM32F4_UsbClient_ClearEndpoints(int32_t endpoint);
-bool STM32F4_UsbClient_RxEnable(USB_CONTROLLER_STATE* usbState, int32_t endpoint);
-
-TinyCLR_UsbClient_DataReceivedHandler STM32F4_UsbClient_DataReceivedHandler;
-TinyCLR_UsbClient_OsExtendedPropertyHandler STM32F4_UsbClient_OsExtendedPropertyHandler;
-
-// usb fifo buffer
-static int32_t usb_fifo_buffer_in[STM32F4_USB_QUEUE_SIZE];
-static int32_t usb_fifo_buffer_out[STM32F4_USB_QUEUE_SIZE];
-static int32_t usb_fifo_buffer_count[STM32F4_USB_QUEUE_SIZE];
-
-static const STM32F4_Gpio_Pin g_STM32F4_Usb_Dm_Pins[] = STM32F4_USB_DM_PINS;
-static const STM32F4_Gpio_Pin g_STM32F4_Usb_Dp_Pins[] = STM32F4_USB_DP_PINS;
-static const STM32F4_Gpio_Pin g_STM32F4_Usb_Vb_Pins[] = STM32F4_USB_VB_PINS;
-static const STM32F4_Gpio_Pin g_STM32F4_Usb_Id_Pins[] = STM32F4_USB_ID_PINS;
-
-static const int32_t TOTAL_USB_CONTROLLERS = SIZEOF_ARRAY(g_STM32F4_Usb_Dm_Pins);
-
-int8_t STM32F4_UsbClient_EndpointMap[] = { ENDPOINT_INUSED_MASK,                          // Endpoint 0
-                                            ENDPOINT_DIR_IN_MASK | ENDPOINT_DIR_OUT_MASK,  // Endpoint 1
-                                            ENDPOINT_DIR_IN_MASK | ENDPOINT_DIR_OUT_MASK,  // Endpoint 2
-                                            ENDPOINT_DIR_IN_MASK | ENDPOINT_DIR_OUT_MASK   // Endpoint 3
-};
-
-USB_CONTROLLER_STATE usbClientState[TOTAL_USB_CONTROLLERS];
-
-uint8_t USB_LanguageDescriptor[USB_LANGUAGE_DESCRIPTOR_SIZE] =
-{
-    USB_LANGUAGE_DESCRIPTOR_SIZE,
-    USB_STRING_DESCRIPTOR_TYPE,
-    0x09, 0x04                      // U.S. English
-};
-
-// Device descriptor
-TinyCLR_UsbClient_DeviceDescriptor deviceDescriptor = {
-
-    {
-        USB_DEVICE_DESCRIPTOR_MARKER,
-        0,
-        sizeof(TinyCLR_UsbClient_DeviceDescriptor)
-    },
-    USB_DEVICE_DESCRIPTOR_LENGTH,       // Length of device descriptor
-    USB_DEVICE_DESCRIPTOR_TYPE,         // USB device descriptor type
-    0x0200,                             // USB Version 2.00 (BCD) (2.0 required for Extended ID recognition)
-    0,                                  // Device class (none)
-    0,                                  // Device subclass (none)
-    0,                                  // Device protocol (none)
-    USB_MAX_EP0_SIZE,                       // Endpoint 0 size
-    USB_DEBUGGER_VENDOR_ID,             // Vendor ID
-    USB_DEBUGGER_PRODUCT_ID,            // Product ID
-    DEVICE_RELEASE_VERSION,             // Product version 1.00 (BCD)
-    MANUFACTURER_NAME_INDEX,            // Manufacturer name string index
-    PRODUCT_NAME_INDEX,                 // Product name string index
-    0,                                  // Serial number string index (none)
-    1                                   // Number of configurations
-};
-
-// configuration descriptor
-TinyCLR_UsbClient_ConfigurationDescriptor configDescriptor = {
-    {
-        USB_CONFIGURATION_DESCRIPTOR_MARKER,
-        0,
-        sizeof(TinyCLR_UsbClient_ConfigurationDescriptor)
-    },
-    USB_CONFIGURATION_DESCRIPTOR_LENGTH,
-    USB_CONFIGURATION_DESCRIPTOR_TYPE,
-    USB_CONFIGURATION_DESCRIPTOR_LENGTH
-        + sizeof(TinyCLR_UsbClient_InterfaceDescriptor)
-        + sizeof(TinyCLR_UsbClient_EndpointDescriptor)
-        + sizeof(TinyCLR_UsbClient_EndpointDescriptor),
-    1,                                                  // Number of interfaces
-    1,                                                  // Number of this configuration
-    0,                                                  // Config descriptor string index (none)
-    (USB_ATTRIBUTE_BASE | USB_ATTRIBUTE_SELF_POWER),    // Config attributes
-    50,                                                 // Device max current draw
-
-    // Interface
-    sizeof(TinyCLR_UsbClient_InterfaceDescriptor),
-    USB_INTERFACE_DESCRIPTOR_TYPE,
-    0,                                          // Interface number
-    0,                                          // Alternate number (main)
-    2,                                          // Number of endpoints
-    0xFF,                                       // Interface class (vendor)
-    1,                                          // Interface subclass
-    1,                                          // Interface protocol
-    0 ,                                          // Interface descriptor string index (none)
-
-    // Endpoint
-    sizeof(TinyCLR_UsbClient_EndpointDescriptor),
-    USB_ENDPOINT_DESCRIPTOR_TYPE,
-    USB_ENDPOINT_DIRECTION_IN,
-    USB_ENDPOINT_ATTRIBUTE_BULK,
-    USB_MAX_EP_SIZE,                                // Endpoint 1 packet size
-    0 ,                                          // Endpoint 1 interval
-
-    //Endpoint
-    sizeof(TinyCLR_UsbClient_EndpointDescriptor),
-    USB_ENDPOINT_DESCRIPTOR_TYPE,
-    USB_ENDPOINT_DIRECTION_OUT,
-    USB_ENDPOINT_ATTRIBUTE_BULK,
-    USB_MAX_EP_SIZE,                                // Endpoint 1 packet size
-    0                                           // Endpoint 1 interval
-};
-
-// Manufacturer name string descriptor header
-TinyCLR_UsbClient_StringDescriptorHeader stringManufacturerDescriptorHeader = {
-        {
-            USB_STRING_DESCRIPTOR_MARKER,
-            MANUFACTURER_NAME_INDEX,
-            sizeof(TinyCLR_UsbClient_StringDescriptorHeader)
-        },
-        USB_STRING_DESCRIPTOR_HEADER_LENGTH + (sizeof(wchar_t) * USB_STRING_DESCRIPTOR_SIZE),
-        USB_STRING_DESCRIPTOR_TYPE,
-        CONCAT(L, DEVICE_MANUFACTURER)
-};
-
-// Product name string descriptor header
-TinyCLR_UsbClient_StringDescriptorHeader stringProductNameDescriptorHeader = {
-    {
-        USB_STRING_DESCRIPTOR_MARKER,
-        PRODUCT_NAME_INDEX,
-        sizeof(TinyCLR_UsbClient_StringDescriptorHeader)
-    },
-    USB_STRING_DESCRIPTOR_HEADER_LENGTH + (sizeof(wchar_t) * USB_STRING_DESCRIPTOR_SIZE),
-    USB_STRING_DESCRIPTOR_TYPE,
-    CONCAT(L, DEVICE_NAME)
-};
-
-// String 4 descriptor header (display name)
-TinyCLR_UsbClient_StringDescriptorHeader stringDisplayNameDescriptorHeader = {
-    {
-        USB_STRING_DESCRIPTOR_MARKER,
-        USB_DISPLAY_STRING_NUM,
-        sizeof(TinyCLR_UsbClient_StringDescriptorHeader)
-    },
-    USB_STRING_DESCRIPTOR_HEADER_LENGTH + (sizeof(wchar_t) * USB_STRING_DESCRIPTOR_SIZE),
-    USB_STRING_DESCRIPTOR_TYPE,
-    CONCAT(L, DEVICE_NAME)
-};
-
-// String 5 descriptor header (friendly name)
-TinyCLR_UsbClient_StringDescriptorHeader stringFriendlyNameDescriptorHeader = {
-    {
-        USB_STRING_DESCRIPTOR_MARKER,
-        USB_FRIENDLY_STRING_NUM,
-        sizeof(TinyCLR_UsbClient_StringDescriptorHeader)
-    },
-    USB_STRING_DESCRIPTOR_HEADER_LENGTH + (sizeof(wchar_t) * USB_STRING_DESCRIPTOR_SIZE),
-    USB_STRING_DESCRIPTOR_TYPE,
-    CONCAT(L, DEVICE_NAME)
-};
-
-// OS Descriptor string for Extended OS Compat ID
-TinyCLR_UsbClient_OsStringDescriptor STM32F4_UsbClient_OsStringDescriptor;
-
-// OS Extended Compatible ID for WinUSB
-TinyCLR_UsbClient_XCompatibleOsId STM32F4_UsbClient_XCompatibleOsId;
-
-// OS Extended Property
-TinyCLR_UsbClient_XPropertiesOsWinUsb STM32F4_UsbClient_XPropertiesOsWinUsb;
-
-// End of configuration marker
-const TinyCLR_UsbClient_DescriptorHeader usbDescriptorHeader = {
-    USB_END_DESCRIPTOR_MARKER,
-    0,
-    0
-};
-
-void STM32F4_UsbClient_SetEvent(USB_CONTROLLER_STATE *usbState, uint32_t event) {
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    uint32_t old_event = usbState->event;
-
-    usbState->event |= event;
-
-    if (old_event != usbState->event) {
-        STM32F4_UsbClient_DataReceivedHandler(nullptr);
-    }
-}
-
-void STM32F4_UsbClient_ClearEvent(USB_CONTROLLER_STATE *usbState, uint32_t event) {
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    usbState->event &= ~event;
-}
-
-void STM32F4_UsbClient_ClearQueues(USB_CONTROLLER_STATE *usbState, bool ClrRxQueue, bool ClrTxQueue) {
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    if (ClrRxQueue) {
-        for (int32_t endpoint = 0; endpoint < STM32F4_USB_QUEUE_SIZE; endpoint++) {
-            if (usbState->queues[endpoint] == nullptr || usbState->isTxQueue[endpoint])
-                continue;
-
-            STM32F4_UsbClient_ClearEndpoints(endpoint);
-
-            /* since this queue is now reset, we have room available for newly arrived packets */
-            STM32F4_UsbClient_RxEnable(usbState, endpoint);
-        }
-    }
-
-    if (ClrTxQueue) {
-        for (int32_t endpoint = 0; endpoint < STM32F4_USB_QUEUE_SIZE; endpoint++) {
-            if (usbState->queues[endpoint] && usbState->isTxQueue[endpoint])
-                STM32F4_UsbClient_ClearEndpoints(endpoint);
-        }
-    }
-}
-
-void STM32F4_UsbClient_StateCallback(USB_CONTROLLER_STATE* usbState) {
-    if (usbState->currentState != usbState->deviceState) {
-        /* whenever we leave the configured state, re-initialize all of the queues */
-        //Not necessary, as TxBuffer may hold any data and then send them out when it is configured again.
-        // The RxQueue is clear when it is configured.
-        if (USB_DEVICE_STATE_CONFIGURED == usbState->currentState) {
-            STM32F4_UsbClient_ClearQueues(usbState, true, true);
-        }
-
-        usbState->currentState = usbState->deviceState;
-
-        switch (usbState->deviceState) {
-        case USB_DEVICE_STATE_DETACHED:
-            usbState->residualCount = 0;
-            usbState->dataCallback = nullptr;
-
-            break;
-        case USB_DEVICE_STATE_CONFIGURED:
-            /* whenever we enter the configured state, re-initialize all of the RxQueues */
-            /* Txqueue has stored some data to be transmitted */
-            STM32F4_UsbClient_ClearQueues(usbState, true, false);
-            break;
-        }
-    }
-}
-
-void STM32F4_UsbClient_DataCallback(USB_CONTROLLER_STATE* usbState) {
-    uint32_t length = __min(usbState->packetSize, usbState->residualCount);
-
-    memcpy(usbState->ptrData, usbState->residualData, length);
-
-    usbState->dataSize = length;
-    usbState->residualData += length;
-    usbState->residualCount -= length;
-
-    if (length == usbState->packetSize) {
-        usbState->expected -= length;
-    }
-    else {
-        usbState->expected = 0;
-    }
-
-    if (usbState->expected) {
-        usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-    }
-    else {
-        usbState->dataCallback = nullptr;
-    }
-}
-
-uint8_t STM32F4_UsbClient_HandleGetStatus(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup) {
-    uint16_t* status;
-    uint16_t  zero = 0;
-
-    /* validate setup packet */
-    if (Setup->wValue != 0 || Setup->wLength != 2) {
-        return USB_STATE_STALL;
-    }
-
-    /* validate based on device state */
-    if (usbState->deviceState == USB_DEVICE_STATE_DEFAULT) {
-        return USB_STATE_STALL;
-    }
-
-    switch (USB_SETUP_RECIPIENT(Setup->bmRequestType)) {
-    case USB_SETUP_RECIPIENT_DEVICE:
-        status = &usbState->deviceStatus;
-        break;
-
-    case USB_SETUP_RECIPIENT_INTERFACE:
-        if (usbState->deviceState != USB_DEVICE_STATE_CONFIGURED) {
-            return USB_STATE_STALL;
-        }
-
-        status = &zero;
-        break;
-
-    case USB_SETUP_RECIPIENT_ENDPOINT:
-        if (usbState->deviceState == USB_DEVICE_STATE_ADDRESS && Setup->wIndex != 0) {
-            return USB_STATE_STALL;
-        }
-
-        /* bit 0x80 designates direction, which we don't utilize in this calculation */
-        Setup->wIndex &= 0x7F;
-
-        if (Setup->wIndex >= usbState->endpointCount) {
-            return USB_STATE_STALL;
-        }
-
-        status = &usbState->endpointStatus[Setup->wIndex];
-        break;
-
-    default:
-        return USB_STATE_STALL;
-    }
-
-    /* send requested status to host */
-    usbState->residualData = (uint8_t*)status;
-    usbState->residualCount = 2;
-    usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-
-    return USB_STATE_DATA;
-}
-
-uint8_t STM32F4_UsbClient_HandleClearFeature(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup) {
-    TinyCLR_UsbClient_ConfigurationDescriptor * Config;
-    uint8_t       retState;
-
-    /* validate setup packet */
-    if (Setup->wLength != 0) {
-        return USB_STATE_STALL;
-    }
-
-    /* validate based on device state */
-    if (usbState->deviceState != USB_DEVICE_STATE_CONFIGURED) {
-        return USB_STATE_STALL;
-    }
-
-    switch (USB_SETUP_RECIPIENT(Setup->bmRequestType)) {
-    case USB_SETUP_RECIPIENT_DEVICE:
-        // only support Remote wakeup
-        if (Setup->wValue != USB_FEATURE_DEVICE_REMOTE_WAKEUP)
-            return USB_STATE_STALL;
-
-        // Locate the configuration descriptor
-        Config = (TinyCLR_UsbClient_ConfigurationDescriptor *)USB_FindRecord(usbState, USB_CONFIGURATION_DESCRIPTOR_MARKER, Setup);
-
-        if (Config && (Config->bmAttributes & USB_ATTRIBUTE_REMOTE_WAKEUP)) {
-            usbState->deviceStatus &= ~USB_STATUS_DEVICE_REMOTE_WAKEUP;
-            retState = USB_STATE_REMOTE_WAKEUP;
-        }
-        else {
-            return USB_STATE_STALL;
-        }
-        break;
-
-    case USB_SETUP_RECIPIENT_INTERFACE:
-        /* there are no interface features to clear */
-        return USB_STATE_STALL;
-
-    case USB_SETUP_RECIPIENT_ENDPOINT:
-        if (usbState->deviceState == USB_DEVICE_STATE_ADDRESS && Setup->wIndex != 0)
-            return USB_STATE_STALL;
-
-        /* bit 0x80 designates direction, which we dont utilize in this calculation */
-        Setup->wIndex &= 0x7F;
-
-        if (Setup->wIndex == 0 || Setup->wIndex >= usbState->endpointCount)
-            return USB_STATE_STALL;
-
-        if (Setup->wValue != USB_FEATURE_ENDPOINT_HALT)
-            return USB_STATE_STALL;
-
-        /* clear the halt feature */
-        usbState->endpointStatus[Setup->wIndex] &= ~USB_STATUS_ENDPOINT_HALT;
-        usbState->endpointStatusChange = Setup->wIndex;
-        retState = USB_STATE_STATUS;
-        break;
-
-    default:
-        return USB_STATE_STALL;
-    }
-
-    /* send zero-length packet to tell host we're done */
-    usbState->residualCount = 0;
-    usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-
-    /* notify lower layer of status change */
-    return retState;
-}
-
-uint8_t STM32F4_UsbClient_HandleSetFeature(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup) {
-    TinyCLR_UsbClient_ConfigurationDescriptor * Config;
-    uint8_t       retState;
-
-    /* validate setup packet */
-    if (Setup->wLength != 0) {
-        return USB_STATE_STALL;
-    }
-
-    /* validate based on device state */
-    if (usbState->deviceState == USB_DEVICE_STATE_DEFAULT) {
-        return USB_STATE_STALL;
-    }
-
-    switch (USB_SETUP_RECIPIENT(Setup->bmRequestType)) {
-    case USB_SETUP_RECIPIENT_DEVICE:
-        // only support Remote wakeup
-        if (Setup->wValue != USB_FEATURE_DEVICE_REMOTE_WAKEUP) {
-            return USB_STATE_STALL;
-        }
-
-        Config = (TinyCLR_UsbClient_ConfigurationDescriptor *)USB_FindRecord(usbState, USB_CONFIGURATION_DESCRIPTOR_MARKER, Setup);
-        if (Config == nullptr)        // If the configuration record could not be found
-            return USB_STATE_STALL; // Something pretty serious is wrong
-
-        if (Config->bmAttributes & USB_ATTRIBUTE_REMOTE_WAKEUP) {
-            usbState->deviceStatus |= USB_STATUS_DEVICE_REMOTE_WAKEUP;
-        }
-
-        retState = USB_STATE_REMOTE_WAKEUP;
-        break;
-
-    case USB_SETUP_RECIPIENT_INTERFACE:
-        /* there are no interface features to set */
-        return USB_STATE_STALL;
-
-    case USB_SETUP_RECIPIENT_ENDPOINT:
-        if (usbState->deviceState == USB_DEVICE_STATE_ADDRESS && Setup->wIndex != 0) {
-            return USB_STATE_STALL;
-        }
-
-        /* bit 0x80 designates direction, which we don't utilize in this calculation */
-        Setup->wIndex &= 0x7F;
-
-        if (Setup->wIndex == 0 || Setup->wIndex >= usbState->endpointCount) {
-            return USB_STATE_STALL;
-        }
-
-        if (Setup->wValue != USB_FEATURE_ENDPOINT_HALT) {
-            return USB_STATE_STALL;
-        }
-
-        /* set the halt feature */
-        usbState->endpointStatus[Setup->wIndex] |= USB_STATUS_ENDPOINT_HALT;
-        usbState->endpointStatusChange = Setup->wIndex;
-        retState = USB_STATE_STATUS;
-        break;
-
-    default:
-        return USB_STATE_STALL;
-    }
-
-    /* send zero-length packet to tell host we're done */
-    usbState->residualCount = 0;
-    usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-
-    /* notify lower layer of status change */
-    return retState;
-}
-
-uint8_t STM32F4_UsbClient_HandleSetAddress(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup) {
-    /* validate setup packet */
-    if (Setup->wValue > 127 || Setup->wIndex != 0 || Setup->wLength != 0) {
-        return USB_STATE_STALL;
-    }
-
-    /* validate based on device state */
-    if (usbState->deviceState >= USB_DEVICE_STATE_CONFIGURED) {
-        return USB_STATE_STALL;
-    }
-
-    /* set address */
-    usbState->address = Setup->wValue;
-
-    /* catch state changes */
-    if (usbState->address == 0) {
-        usbState->deviceState = USB_DEVICE_STATE_DEFAULT;
-    }
-    else {
-        usbState->deviceState = USB_DEVICE_STATE_ADDRESS;
-    }
-
-    STM32F4_UsbClient_StateCallback(usbState);
-
-    /* send zero-length packet to tell host we're done */
-    usbState->residualCount = 0;
-    usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-
-    /* notify hardware of address change */
-    return USB_STATE_ADDRESS;
-}
-
-uint8_t STM32F4_UsbClient_HandleConfigurationRequests(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup) {
-    const TinyCLR_UsbClient_DescriptorHeader * header;
-    uint8_t       type;
-    uint8_t       DescriptorIndex;
-
-    /* this request is valid regardless of device state */
-    type = ((Setup->wValue & 0xFF00) >> 8);
-    DescriptorIndex = (Setup->wValue & 0x00FF);
-    usbState->expected = Setup->wLength;
-
-    if (usbState->expected == 0) {
-        // just return an empty Status packet
-        usbState->residualCount = 0;
-        usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-        return USB_STATE_DATA;
-    }
-
-    // The very first GET_DESCRIPTOR command out of reset should always return at most packetSize bytes.
-    // After that, you can return as many as the host has asked.
-    if (usbState->deviceState <= USB_DEVICE_STATE_DEFAULT) {
-        if (usbState->firstGetDescriptor) {
-            usbState->firstGetDescriptor = false;
-
-            usbState->expected = __min(usbState->expected, usbState->packetSize);
-        }
-    }
-
-    usbState->residualData = nullptr;
-    usbState->residualCount = 0;
-
-    if (Setup->bRequest == USB_GET_DESCRIPTOR) {
-        switch (type) {
-        case USB_DEVICE_DESCRIPTOR_TYPE:
-            header = USB_FindRecord(usbState, USB_DEVICE_DESCRIPTOR_MARKER, Setup);
-            if (header) {
-                const TinyCLR_UsbClient_DeviceDescriptor * device = (TinyCLR_UsbClient_DeviceDescriptor *)header;
-                usbState->residualData = (uint8_t *)&device->bLength;      // Start of the device descriptor
-                usbState->residualCount = __min(usbState->expected, device->bLength);
-            }
-            break;
-
-        case USB_CONFIGURATION_DESCRIPTOR_TYPE:
-            header = USB_FindRecord(usbState, USB_CONFIGURATION_DESCRIPTOR_MARKER, Setup);
-            if (header) {
-                const TinyCLR_UsbClient_ConfigurationDescriptor * Config = (TinyCLR_UsbClient_ConfigurationDescriptor *)header;
-                usbState->residualData = (uint8_t *)&Config->bLength;
-                usbState->residualCount = __min(usbState->expected, Config->wTotalLength);
-            }
-            break;
-
-        case USB_STRING_DESCRIPTOR_TYPE:
-            if (DescriptorIndex == 0)        // If host is requesting the language list
-            {
-                usbState->residualData = USB_LanguageDescriptor;
-                usbState->residualCount = __min(usbState->expected, USB_LANGUAGE_DESCRIPTOR_SIZE);
-            }
-            else if (nullptr != (header = USB_FindRecord(usbState, USB_STRING_DESCRIPTOR_MARKER, Setup))) {
-                const TinyCLR_UsbClient_StringDescriptorHeader * string = (TinyCLR_UsbClient_StringDescriptorHeader *)header;
-                usbState->residualData = (uint8_t *)&string->bLength;
-                usbState->residualCount = __min(usbState->expected, string->bLength);
-            }
-            break;
-
-        default:
-            break;
-        }
-    }
-
-    // If the request was not recognized, the generic types should be searched
-    if (usbState->residualData == nullptr) {
-        if (nullptr != (header = USB_FindRecord(usbState, USB_GENERIC_DESCRIPTOR_MARKER, Setup))) {
-            usbState->residualData = (uint8_t *)header;
-            usbState->residualData += sizeof(TinyCLR_UsbClient_GenericDescriptorHeader);       // ptrData is located right after the header
-            usbState->residualCount = __min(usbState->expected, header->size - sizeof(TinyCLR_UsbClient_GenericDescriptorHeader));
-        }
-        else
-            return USB_STATE_STALL;
-    }
-
-    usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-
-    return USB_STATE_DATA;
-}
-
-uint8_t STM32F4_UsbClient_HandleGetConfiguration(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup) {
-    /* validate setup packet */
-    if (Setup->wValue != 0 || Setup->wIndex != 0 || Setup->wLength != 1) {
-        return USB_STATE_STALL;
-    }
-
-    /* validate based on device state */
-    if (usbState->deviceState == USB_DEVICE_STATE_DEFAULT) {
-        return USB_STATE_STALL;
-    }
-
-    usbState->residualData = &usbState->configurationNum;
-    usbState->residualCount = 1;
-    usbState->expected = 1;
-    usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-
-    return USB_STATE_DATA;
-}
-
-uint8_t STM32F4_UsbClient_HandleSetConfiguration(USB_CONTROLLER_STATE* usbState, USB_SETUP_PACKET* Setup, bool DataPhase) {
-    /* validate setup packet */
-    if (Setup->wIndex != 0 || Setup->wLength != 0) {
-        return USB_STATE_STALL;
-    }
-
-    /* validate based on device state */
-    if (usbState->deviceState == USB_DEVICE_STATE_DEFAULT) {
-        return USB_STATE_STALL;
-    }
-
-    /* we only support one configuration */
-    if (Setup->wValue > 1) {
-        return USB_STATE_STALL;
-    }
-
-    usbState->configurationNum = Setup->wValue;
-
-    /* catch state changes */
-    if (usbState->configurationNum == 0) {
-        usbState->deviceState = USB_DEVICE_STATE_ADDRESS;
-    }
-    else {
-        usbState->deviceState = USB_DEVICE_STATE_CONFIGURED;
-    }
-
-    STM32F4_UsbClient_StateCallback(usbState);
-
-    if (DataPhase) {
-        /* send zero-length packet to tell host we're done */
-        usbState->residualCount = 0;
-        usbState->dataCallback = STM32F4_UsbClient_DataCallback;
-    }
-
-    return USB_STATE_CONFIGURATION;
-}
-
-// Searches through the USB configuration records for the requested type
-// Returns a pointer to the header information if found and nullptr if not
-const TinyCLR_UsbClient_DescriptorHeader * USB_FindRecord(USB_CONTROLLER_STATE* usbState, uint8_t marker, USB_SETUP_PACKET * setup) {
-    bool Done = false;
-
-    const TinyCLR_UsbClient_DescriptorHeader * header = (const TinyCLR_UsbClient_DescriptorHeader *)usbState->configuration;
-    TinyCLR_UsbClient_DescriptorHeader* ptr = (TinyCLR_UsbClient_DescriptorHeader*)(*(uint32_t*)header);
-    TinyCLR_UsbClient_ConfigurationDescriptor *config;
-
-    // If there is no configuration for this controller
-    if (nullptr == header)
-        return header;
-
-    while (!Done) {
-        const uint8_t *next = (const uint8_t *)header;
-        ptr = (TinyCLR_UsbClient_DescriptorHeader*)(*(uint32_t*)header);
-        next += 4;      // Calculate address of next record
-
-        const TinyCLR_UsbClient_GenericDescriptorHeader *generic = (TinyCLR_UsbClient_GenericDescriptorHeader *)ptr;
-
-        switch (ptr->marker) {
-        case USB_DEVICE_DESCRIPTOR_MARKER:
-            if (ptr->marker == marker)
-                Done = true;
-            break;
-
-        case USB_CONFIGURATION_DESCRIPTOR_MARKER:
-            config = (TinyCLR_UsbClient_ConfigurationDescriptor *)UsbDefaultConfiguration.config;
-            if (config->header.marker == marker)
-                Done = true;
-            break;
-
-        case USB_STRING_DESCRIPTOR_MARKER:
-            // If String descriptor then the index is significant
-            if ((ptr->marker == marker) && (ptr->iValue == (setup->wValue & 0x00FF)))
-                Done = true;
-            break;
-        case USB_GENERIC_DESCRIPTOR_MARKER:
-            if (generic->bmRequestType == setup->bmRequestType &&
-                generic->bRequest == setup->bRequest &&
-                generic->wValue == setup->wValue &&
-                generic->wIndex == setup->wIndex) {
-                Done = true;
-            }
-            break;
-        case USB_END_DESCRIPTOR_MARKER:
-            Done = true;
-            header = nullptr;
-            ptr = nullptr;
-            break;
-        }
-        if (!Done)
-            header = (const TinyCLR_UsbClient_DescriptorHeader *)next;    // Try next record
-    }
-
-    return ptr;
-}
-
-uint8_t STM32F4_UsbClient_ControlCallback(USB_CONTROLLER_STATE* usbState) {
-    USB_SETUP_PACKET* Setup;
-
-    if (usbState->dataSize == 0) {
-        return USB_STATE_DONE;
-    }
-
-    Setup = (USB_SETUP_PACKET*)usbState->ptrData;
-
-    switch (Setup->bRequest) {
-    case USB_GET_STATUS:
-        return STM32F4_UsbClient_HandleGetStatus(usbState, Setup);
-    case USB_CLEAR_FEATURE:
-        return STM32F4_UsbClient_HandleClearFeature(usbState, Setup);
-    case USB_SET_FEATURE:
-        return STM32F4_UsbClient_HandleSetFeature(usbState, Setup);
-    case USB_SET_ADDRESS:
-        return STM32F4_UsbClient_HandleSetAddress(usbState, Setup);
-    case USB_GET_CONFIGURATION:
-        return STM32F4_UsbClient_HandleGetConfiguration(usbState, Setup);
-    case USB_SET_CONFIGURATION:
-        return STM32F4_UsbClient_HandleSetConfiguration(usbState, Setup, true);
-    default:
-        return STM32F4_UsbClient_HandleConfigurationRequests(usbState, Setup);
-    }
-
-    return USB_STATE_STALL;
-}
-
-USB_PACKET64* STM32F4_UsbClient_RxEnqueue(USB_CONTROLLER_STATE* usbState, int32_t endpoint, bool& disableRx) {
-    USB_PACKET64* packet;
-
-    if (usb_fifo_buffer_count[endpoint] == STM32F4_USB_FIFO_BUFFER_SIZE) {
-        disableRx = true;
-
-        return nullptr;
-    }
-
-    disableRx = false;
-
-    packet = &usbState->queues[endpoint][usb_fifo_buffer_in[endpoint]];
-
-    usb_fifo_buffer_in[endpoint]++;
-    usb_fifo_buffer_count[endpoint]++;
-
-    if (usb_fifo_buffer_in[endpoint] == STM32F4_USB_FIFO_BUFFER_SIZE)
-        usb_fifo_buffer_in[endpoint] = 0;
-
-    STM32F4_UsbClient_SetEvent(usbState, 1 << endpoint);
-
-    return packet;
-}
-
-USB_PACKET64* STM32F4_UsbClient_TxDequeue(USB_CONTROLLER_STATE* usbState, int32_t endpoint) {
-    USB_PACKET64* packet;
-
-    if (usb_fifo_buffer_count[endpoint] == 0) {
-        return nullptr;
-    }
-
-    packet = &usbState->queues[endpoint][usb_fifo_buffer_out[endpoint]];
-
-    usb_fifo_buffer_count[endpoint]--;
-    usb_fifo_buffer_out[endpoint]++;
-
-    if (usb_fifo_buffer_out[endpoint] == STM32F4_USB_FIFO_BUFFER_SIZE)
-        usb_fifo_buffer_out[endpoint] = 0;
-
-    return packet;
-}
-
-void STM32F4_UsbClient_ClearEndpoints(int32_t endpoint) {
-    usb_fifo_buffer_in[endpoint] = usb_fifo_buffer_out[endpoint] = usb_fifo_buffer_count[endpoint] = 0;
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-/// TinyCLR USBClient API
-///////////////////////////////////////////////////////////////////////////////////////////
-bool STM32F4_UsbClient_Initialize(USB_CONTROLLER_STATE* usbState);
-bool STM32F4_UsbClient_Uninitialize(USB_CONTROLLER_STATE* usbState);
-bool STM32F4_UsbClient_StartOutput(USB_CONTROLLER_STATE* usbState, int32_t endpoint);
-
-static TinyCLR_UsbClient_Provider usbClientProvider;
-static TinyCLR_Api_Info usbClientApi;
-
-const TinyCLR_Api_Info* STM32F4_UsbClient_GetApi() {
-    usbClientProvider.Parent = &usbClientApi;
-    usbClientProvider.Index = 0;
-    usbClientProvider.Acquire = &STM32F4_UsbClient_Acquire;
-    usbClientProvider.Release = &STM32F4_UsbClient_Release;
-    usbClientProvider.Open = &STM32F4_UsbClient_Open;
-    usbClientProvider.Close = &STM32F4_UsbClient_Close;
-    usbClientProvider.Write = &STM32F4_UsbClient_Write;
-    usbClientProvider.Read = &STM32F4_UsbClient_Read;
-    usbClientProvider.Flush = &STM32F4_UsbClient_Flush;
-    usbClientProvider.SetDeviceDescriptor = &STM32F4_UsbClient_SetDeviceDescriptor;
-    usbClientProvider.SetConfigDescriptor = &STM32F4_UsbClient_SetConfigDescriptor;
-    usbClientProvider.SetStringDescriptor = &STM32F4_UsbClient_SetStringDescriptor;
-    usbClientProvider.SetDataReceivedHandler = &STM32F4_UsbClient_SetDataReceivedHandler;
-    usbClientProvider.SetOsExtendedPropertyHandler = &STM32F4_UsbClient_SetOsExtendedPropertyHandler;
-
-    usbClientApi.Author = "GHI Electronics, LLC";
-    usbClientApi.Name = "GHIElectronics.TinyCLR.NativeApis.STM32F4.UsbClientProvider";
-    usbClientApi.Type = TinyCLR_Api_Type::UsbClientProvider;
-    usbClientApi.Version = 0;
-    usbClientApi.Count = 1;
-    usbClientApi.Implementation = &usbClientProvider;
-
-    return &usbClientApi;
-}
-
-TinyCLR_Result STM32F4_UsbClient_Acquire(const TinyCLR_UsbClient_Provider* self) {
-    int32_t controller = self->Index;
-
-    uint8_t *osStringDescriptor = (uint8_t*)&STM32F4_UsbClient_OsStringDescriptor;
-    uint8_t *xCompatibleOsId = (uint8_t*)&STM32F4_UsbClient_XCompatibleOsId;
-    uint8_t *xPropertiesOsWinUsb = (uint8_t*)&STM32F4_UsbClient_XPropertiesOsWinUsb;
-
-    STM32F4_UsbClient_OsExtendedPropertyHandler(self, osStringDescriptor, xCompatibleOsId, xPropertiesOsWinUsb);
-
-    USB_CONTROLLER_STATE *usbState = &usbClientState[controller];
-
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    // Init UsbDefaultConfiguration
-    memset(&UsbDefaultConfiguration, 0, sizeof(USB_DYNAMIC_CONFIGURATION));
-
-    configDescriptor.epWrite.bEndpointAddress = USB_ENDPOINT_DIRECTION_IN;
-    configDescriptor.epRead.bEndpointAddress = USB_ENDPOINT_DIRECTION_OUT;
-
-    UsbDefaultConfiguration.device = (TinyCLR_UsbClient_DeviceDescriptor*)&deviceDescriptor;
-    UsbDefaultConfiguration.config = (TinyCLR_UsbClient_ConfigurationDescriptor*)&configDescriptor;
-
-    UsbDefaultConfiguration.manHeader = (TinyCLR_UsbClient_StringDescriptorHeader*)&stringManufacturerDescriptorHeader;
-    UsbDefaultConfiguration.prodHeader = (TinyCLR_UsbClient_StringDescriptorHeader*)&stringProductNameDescriptorHeader;
-    UsbDefaultConfiguration.displayStringHeader = (TinyCLR_UsbClient_StringDescriptorHeader*)&stringDisplayNameDescriptorHeader;
-    UsbDefaultConfiguration.friendlyStringHeader = (TinyCLR_UsbClient_StringDescriptorHeader*)&stringFriendlyNameDescriptorHeader;
-
-    UsbDefaultConfiguration.OS_String = (TinyCLR_UsbClient_OsStringDescriptor*)&STM32F4_UsbClient_OsStringDescriptor;
-    UsbDefaultConfiguration.OS_XCompatible_ID = (TinyCLR_UsbClient_XCompatibleOsId*)&STM32F4_UsbClient_XCompatibleOsId;
-    UsbDefaultConfiguration.OS_XProperty = (TinyCLR_UsbClient_XPropertiesOsWinUsb*)&STM32F4_UsbClient_XPropertiesOsWinUsb;
-
-    UsbDefaultConfiguration.endList = (TinyCLR_UsbClient_DescriptorHeader*)&usbDescriptorHeader;
-
-    // Init Usb usbState
-    memset(usbState, 0, sizeof(USB_CONTROLLER_STATE));
-
-    usbState->controllerNum = controller;
-    usbState->configuration = &UsbDefaultConfiguration;
-    usbState->currentState = USB_DEVICE_STATE_UNINITIALIZED;
-    usbState->deviceStatus = USB_STATUS_DEVICE_SELF_POWERED;
-    usbState->endpointCount = USB_FS_MAX_EP_COUNT;
-    usbState->packetSize = USB_MAX_EP0_SIZE;
-    usbState->initialized = true;
-
-    for (auto i = 0; i < STM32F4_USB_QUEUE_SIZE; i++) {
-        usbState->pipes[i].RxEP = USB_ENDPOINT_NULL;
-        usbState->pipes[i].TxEP = USB_ENDPOINT_NULL;
-        usbState->maxPacketSize[i] = USB_MAX_EP_SIZE;
-    }
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_Release(const TinyCLR_UsbClient_Provider* self) {
-    int32_t controller = self->Index;
-
-    USB_CONTROLLER_STATE *usbState = &usbClientState[controller];
-
-    if (usbState->initialized) {
-        DISABLE_INTERRUPTS_SCOPED(irq);
-
-        STM32F4_UsbClient_Uninitialize(usbState);
-
-        usbState->initialized = false;
-
-        // for soft reboot allow the USB to be off for at least 100ms
-        STM32F4_Time_Delay(nullptr, 100000); // 100ms
-    }
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_Open(const TinyCLR_UsbClient_Provider* self, int32_t& pipe, TinyCLR_UsbClient_PipeMode mode) {
-    int32_t controller = self->Index;
-
-    USB_CONTROLLER_STATE * usbState = &usbClientState[controller];
-
-    if (!usbState->initialized)     // If no such controller exists (or it is not initialized)
-        return TinyCLR_Result::NotAvailable;
-
-    int32_t writeEp = USB_ENDPOINT_NULL;
-    int32_t readEp = USB_ENDPOINT_NULL;
-
-    if (mode != TinyCLR_UsbClient_PipeMode::InOut)
-        return TinyCLR_Result::NotSupported;
-
-    for (auto i = 0; i < SIZEOF_ARRAY(STM32F4_UsbClient_EndpointMap); i++) {
-        if ((STM32F4_UsbClient_EndpointMap[i] & ENDPOINT_INUSED_MASK)) // in used
-            continue;
-
-        if (writeEp == USB_ENDPOINT_NULL && ((STM32F4_UsbClient_EndpointMap[i] & ENDPOINT_DIR_IN_MASK) == ENDPOINT_DIR_IN_MASK)) {
-            writeEp = i;
-            STM32F4_UsbClient_EndpointMap[i] |= ENDPOINT_INUSED_MASK;
-
-            continue;
-        }
-
-        if (readEp == USB_ENDPOINT_NULL && ((STM32F4_UsbClient_EndpointMap[i] & ENDPOINT_DIR_OUT_MASK) == ENDPOINT_DIR_OUT_MASK)) {
-            readEp = i;
-            STM32F4_UsbClient_EndpointMap[i] |= ENDPOINT_INUSED_MASK;
-
-            continue;
-        }
-
-        if (writeEp != 0 && readEp != 0) {
-            break;
-        }
-    }
-    // Check the usbPipe and the two endpoint numbers for validity (both endpoints cannot be zero)
-    if ((readEp == USB_ENDPOINT_NULL && writeEp == USB_ENDPOINT_NULL)
-        || (readEp != USB_ENDPOINT_NULL && (readEp < 1 || readEp >= STM32F4_USB_QUEUE_SIZE))
-        || (writeEp != USB_ENDPOINT_NULL && (writeEp < 1 || writeEp >= STM32F4_USB_QUEUE_SIZE)))
-        return TinyCLR_Result::NotAvailable;
-
-    // The specified endpoints must not be in use by another pipe
-    for (int32_t i = 0; i < STM32F4_USB_QUEUE_SIZE; i++) {
-        if (readEp != USB_ENDPOINT_NULL && (usbState->pipes[i].RxEP == readEp || usbState->pipes[i].TxEP == readEp))
-            return TinyCLR_Result::NotAvailable;
-        if (writeEp != USB_ENDPOINT_NULL && (usbState->pipes[i].RxEP == writeEp || usbState->pipes[i].TxEP == writeEp))
-            return TinyCLR_Result::NotAvailable;
-    }
-
-    for (pipe = 0; pipe < STM32F4_USB_QUEUE_SIZE; pipe++) {
-        // The Pipe must be currently closed
-        if (usbState->pipes[pipe].RxEP == USB_ENDPOINT_NULL && usbState->pipes[pipe].TxEP == USB_ENDPOINT_NULL)
-            break;
-    }
-
-    if (pipe == STM32F4_USB_QUEUE_SIZE)
-        return TinyCLR_Result::NotAvailable;
-
-    // All tests pass, assign the endpoints to the pipe
-    usbState->pipes[pipe].RxEP = readEp;
-    usbState->pipes[pipe].TxEP = writeEp;
-
-    TinyCLR_UsbClient_ConfigurationDescriptor *config = (TinyCLR_UsbClient_ConfigurationDescriptor *)UsbDefaultConfiguration.config;
-    TinyCLR_UsbClient_EndpointDescriptor  *ep = (TinyCLR_UsbClient_EndpointDescriptor  *)(((uint8_t *)config) + USB_CONFIGURATION_DESCRIPTOR_LENGTH + sizeof(TinyCLR_UsbClient_DescriptorHeader) + sizeof(TinyCLR_UsbClient_InterfaceDescriptor));
-
-    uint8_t * end = ((uint8_t *)config) + config->header.size;
-    uint32_t epType = usbState->endpointType;
-
-    while (((uint8_t *)ep) != nullptr && ((uint8_t *)ep) < end) {
-        if (USB_ENDPOINT_DESCRIPTOR_TYPE != ep->bDescriptorType || sizeof(TinyCLR_UsbClient_EndpointDescriptor) != ep->bLength)
-            break;
-
-        auto idx = 0;
-
-        if (ep->bEndpointAddress == USB_ENDPOINT_DIRECTION_IN) {
-            ep->bEndpointAddress |= writeEp;
-            idx = writeEp;
-            usbState->isTxQueue[idx] = true;
-        }
-
-        else if (ep->bEndpointAddress == USB_ENDPOINT_DIRECTION_OUT) {
-            ep->bEndpointAddress |= readEp;
-            idx = readEp;
-            usbState->isTxQueue[idx] = false;
-        }
-
-        if (idx > 0) {
-            if (apiProvider != nullptr) {
-                auto memoryProvider = (const TinyCLR_Memory_Provider*)apiProvider->FindDefault(apiProvider, TinyCLR_Api_Type::MemoryProvider);
-
-                usbState->queues[idx] = (USB_PACKET64*)memoryProvider->Allocate(memoryProvider, STM32F4_USB_FIFO_BUFFER_SIZE * sizeof(USB_PACKET64));
-
-                if (usbState->queues[idx] == nullptr)
-                    return TinyCLR_Result::ArgumentNull;;
-
-                memset(reinterpret_cast<uint8_t*>(usbState->queues[idx]), 0x00, STM32F4_USB_FIFO_BUFFER_SIZE * sizeof(USB_PACKET64));
-            }
-
-            STM32F4_UsbClient_ClearEndpoints(idx);
-
-            epType |= (ep->bmAttributes & 3) << (idx * 2);
-            usbState->maxPacketSize[idx] = ep->wMaxPacketSize;
-        }
-
-        ep = (TinyCLR_UsbClient_EndpointDescriptor  *)(((uint8_t *)ep) + ep->bLength);
-    }
-
-    usbState->endpointType = epType;
-
-    if (usbState->currentState == USB_DEVICE_STATE_UNINITIALIZED) {
-        STM32F4_UsbClient_Initialize(usbState);
-    }
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_Close(const TinyCLR_UsbClient_Provider* self, int32_t pipe) {
-    int32_t controller = self->Index;
-
-    USB_CONTROLLER_STATE * usbState = &usbClientState[controller];
-
-    if (!usbState->initialized || pipe >= STM32F4_USB_QUEUE_SIZE)
-        return TinyCLR_Result::NotAvailable;
-
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    // Close the Rx pipe
-    int32_t endpoint = usbState->pipes[pipe].RxEP;
-    if (endpoint != USB_ENDPOINT_NULL && usbState->queues[endpoint]) {
-        STM32F4_UsbClient_ClearEndpoints(endpoint);
-    }
-
-    usbState->pipes[pipe].RxEP = USB_ENDPOINT_NULL;
-    //Free endpoint
-    STM32F4_UsbClient_EndpointMap[endpoint] &= ~ENDPOINT_INUSED_MASK;
-
-    // Close the TX pipe
-    endpoint = usbState->pipes[pipe].TxEP;
-    if (endpoint != USB_ENDPOINT_NULL && usbState->queues[endpoint] != nullptr) {
-        STM32F4_UsbClient_ClearEndpoints(endpoint);
-
-        if (apiProvider != nullptr) {
-            auto memoryProvider = (const TinyCLR_Memory_Provider*)apiProvider->FindDefault(apiProvider, TinyCLR_Api_Type::MemoryProvider);
-
-            if (usbState->queues[endpoint] != nullptr)
-                memoryProvider->Free(memoryProvider, usbState->queues[endpoint]);
-
-            usbState->queues[endpoint] = nullptr;
-        }
-    }
-
-    usbState->pipes[pipe].TxEP = USB_ENDPOINT_NULL;
-
-    //Free endpoint
-    STM32F4_UsbClient_EndpointMap[endpoint] &= ~ENDPOINT_INUSED_MASK;
-
-    configDescriptor.epWrite.bEndpointAddress = USB_ENDPOINT_DIRECTION_IN;
-    configDescriptor.epRead.bEndpointAddress = USB_ENDPOINT_DIRECTION_OUT;
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_Write(const TinyCLR_UsbClient_Provider* self, int32_t pipe, const uint8_t* data, size_t& length) {
-    int32_t controller = self->Index;
-
-    USB_CONTROLLER_STATE * usbState = &usbClientState[controller];
-
-    if (pipe >= STM32F4_USB_QUEUE_SIZE
-        || data == nullptr
-        || usbState->deviceState != USB_DEVICE_STATE_CONFIGURED
-        || length == 0) {
-        return TinyCLR_Result::ArgumentInvalid;
-    }
-
-    int32_t endpoint = usbState->pipes[pipe].TxEP;
-    // If no Write side to pipe (or if not yet open)
-    if (endpoint == USB_ENDPOINT_NULL || usbState->queues[endpoint] == nullptr) {
-        return TinyCLR_Result::NotAvailable;
-    }
-
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    const uint8_t*      ptr = data;
-    uint32_t            count = length;
-    bool                Done = false;
-    uint32_t            WaitLoopCnt = 0;
-    int32_t                 totWrite = 0;
-
-    // This loop packetizes the data and sends it out.  All packets sent have
-    // the maximum length for the given endpoint except for the last packet which
-    // will always have less than the maximum length - even if the packet length
-    // must be zero for this to occur.   This is done to comply with standard
-    // USB bulk-mode transfers.
-    while (!Done) {
-
-        USB_PACKET64* Packet64 = nullptr;
-
-        if (usb_fifo_buffer_count[endpoint] < STM32F4_USB_FIFO_BUFFER_SIZE) {
-            Packet64 = &usbState->queues[endpoint][usb_fifo_buffer_in[endpoint]];
-
-            usb_fifo_buffer_in[endpoint]++;
-            usb_fifo_buffer_count[endpoint]++;
-
-            if (usb_fifo_buffer_in[endpoint] == STM32F4_USB_FIFO_BUFFER_SIZE)
-                usb_fifo_buffer_in[endpoint] = 0;
-        }
-
-        if (Packet64) {
-            uint32_t max_move;
-
-            if (count > usbState->maxPacketSize[endpoint])
-                max_move = usbState->maxPacketSize[endpoint];
-            else
-                max_move = count;
-
-            if (max_move) {
-                memcpy(Packet64->Buffer, ptr, max_move);
-            }
-
-            // we are done when we send a non-full length packet
-            if (max_move < usbState->maxPacketSize[endpoint]) {
-                Done = true;
-            }
-
-            Packet64->Size = max_move;
-            count -= max_move;
-            ptr += max_move;
-
-            totWrite += max_move;
-
-            WaitLoopCnt = 0;
-        }
-        if (Packet64 == nullptr) {
-            // a 64-byte USB packet takes less than 50uSec
-            // according to the timing calculations of the USB Chief
-            // this is way too short to bother with a call
-            // to WaitForEventsInternal, so just uSec delay the path
-            // here for 50uSec.
-
-            // if in ISR, return
-
-            // if more than 100*50us=5ms,still no packet avaialable, PC side go wrong,stop the loop
-            // otherwise it will spin here forever and stopwatch get kick in.
-            WaitLoopCnt++;
-            if (WaitLoopCnt > 100) {
-                // if we were unable to send any data then no one is listening so lets
-                if (count == length) {
-                    STM32F4_UsbClient_ClearEndpoints(endpoint);
-                }
-
-                goto done_write;
-            }
-
-            if (irq.IsDisabled()) // @todo - this really needs more checks to be totally valid
-            {
-                goto done_write;
-            }
-
-            if (usbState->deviceState != USB_DEVICE_STATE_CONFIGURED) {
-                goto done_write;
-            }
-
-            STM32F4_UsbClient_StartOutput(usbState, endpoint);
-
-            irq.Release();
-
-            STM32F4_Time_Delay(nullptr, 50);
-
-            irq.Acquire();
-        }
-    }
-
-    // here we have a post-condition that IRQs are disabled for all paths through conditional block above
-    if (usbState->deviceState == USB_DEVICE_STATE_CONFIGURED) {
-        STM32F4_UsbClient_StartOutput(usbState, endpoint);
-    }
-
-done_write:
-    length = totWrite;
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_Read(const TinyCLR_UsbClient_Provider* self, int32_t pipe, uint8_t* data, size_t& length) {
-    int32_t controller = self->Index;
-
-    int32_t endpoint;
-    USB_CONTROLLER_STATE * usbState = &usbClientState[controller];
-
-    if (pipe >= STM32F4_USB_QUEUE_SIZE
-        || usbState->deviceState != USB_DEVICE_STATE_CONFIGURED) {
-        return TinyCLR_Result::ArgumentInvalid;
-    }
-
-    endpoint = usbState->pipes[pipe].RxEP;
-    // If no Read side to pipe (or if not yet open)
-    if (endpoint == USB_ENDPOINT_NULL || usbState->queues[endpoint] == nullptr) {
-        return TinyCLR_Result::NotAvailable;
-    }
-
-    DISABLE_INTERRUPTS_SCOPED(irq);
-
-    USB_PACKET64* Packet64 = nullptr;
-    uint8_t*        ptr = (uint8_t*)data;
-    uint32_t        count = 0;
-    uint32_t        remain = length;
-
-    while (count < length) {
-        uint32_t max_move;
-
-        if (usb_fifo_buffer_count[endpoint] > 0) {
-            Packet64 = &usbState->queues[endpoint][usb_fifo_buffer_out[endpoint]];
-
-            usb_fifo_buffer_count[endpoint]--;
-            usb_fifo_buffer_out[endpoint]++;
-
-            if (usb_fifo_buffer_out[endpoint] == STM32F4_USB_FIFO_BUFFER_SIZE) {
-                usb_fifo_buffer_out[endpoint] = 0;
-            }
-        }
-
-        if (!Packet64) {
-            STM32F4_UsbClient_ClearEvent(usbState, 1 << endpoint);
-            break;
-        }
-
-        max_move = Packet64->Size - usbState->currentPacketOffset[endpoint];
-        if (remain < max_move) max_move = remain;
-
-        memcpy(ptr, &Packet64->Buffer[usbState->currentPacketOffset[endpoint]], max_move);
-
-        usbState->currentPacketOffset[endpoint] += max_move;
-        ptr += max_move;
-        count += max_move;
-        remain -= max_move;
-
-        /* if we're done with this packet, move onto the next */
-        if (usbState->currentPacketOffset[endpoint] == Packet64->Size) {
-            usbState->currentPacketOffset[endpoint] = 0;
-            Packet64 = nullptr;
-
-            STM32F4_UsbClient_RxEnable(usbState, endpoint);
-        }
-    }
-
-    length = count;
-
-    return TinyCLR_Result::Success;
-}
-
-#define USB_FLUSH_RETRY_COUNT 30
-TinyCLR_Result STM32F4_UsbClient_Flush(const TinyCLR_UsbClient_Provider* self, int32_t pipe) {
-    int32_t controller = self->Index;
-
-    int32_t endpoint;
-    int32_t retries = USB_FLUSH_RETRY_COUNT;
-    int32_t queueCnt;
-    USB_CONTROLLER_STATE * usbState = &usbClientState[controller];
-
-    if (pipe >= STM32F4_USB_QUEUE_SIZE) {
-        return TinyCLR_Result::ArgumentInvalid;
-    }
-
-    /* not configured, no data can go in or out */
-    if (usbState->deviceState != USB_DEVICE_STATE_CONFIGURED) {
-        return TinyCLR_Result::InvalidOperation;
-    }
-
-    endpoint = usbState->pipes[pipe].TxEP;
-    // If no Write side to pipe (or if not yet open)
-    if (endpoint == USB_ENDPOINT_NULL || usbState->queues[endpoint] == nullptr) {
-        return TinyCLR_Result::NotAvailable;
-    }
-
-    queueCnt = usb_fifo_buffer_count[endpoint];
-
-    // interrupts were disabled or USB interrupt was disabled for whatever reason, so force the flush
-    while (usb_fifo_buffer_count[endpoint] > 0 && retries > 0) {
-        STM32F4_UsbClient_StartOutput(usbState, endpoint);
-
-        STM32F4_Time_Delay(nullptr, queueCnt == usb_fifo_buffer_count[endpoint] ? 100 : 0); // don't call Events_WaitForEventsXXX because it will turn off interrupts
-
-        retries = (queueCnt == usb_fifo_buffer_count[endpoint]) ? retries - 1 : USB_FLUSH_RETRY_COUNT;
-
-        queueCnt = usb_fifo_buffer_count[endpoint];
-    }
-
-    if (retries <= 0)
-        STM32F4_UsbClient_ClearEndpoints(endpoint);
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_SetDataReceivedHandler(const TinyCLR_UsbClient_Provider* self, TinyCLR_UsbClient_DataReceivedHandler handler) {
-    int32_t controller = self->Index;
-
-    STM32F4_UsbClient_DataReceivedHandler = handler;
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_SetOsExtendedPropertyHandler(const TinyCLR_UsbClient_Provider* self, TinyCLR_UsbClient_OsExtendedPropertyHandler handler) {
-    int32_t controller = self->Index;
-
-    STM32F4_UsbClient_OsExtendedPropertyHandler = handler;
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_SetDeviceDescriptor(const TinyCLR_UsbClient_Provider* self, const void* descriptor, int32_t length) {
-    memcpy(&deviceDescriptor, descriptor, length);
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_SetConfigDescriptor(const TinyCLR_UsbClient_Provider* self, const void* descriptor, int32_t length) {
-    memcpy(&configDescriptor, descriptor, length);
-
-    return TinyCLR_Result::Success;
-}
-
-TinyCLR_Result STM32F4_UsbClient_SetStringDescriptor(const TinyCLR_UsbClient_Provider* self, TinyCLR_UsbClient_StringDescriptorType type, const wchar_t* value) {
-    switch (type) {
-    case TinyCLR_UsbClient_StringDescriptorType::ManufacturerName:
-        memcpy(&stringManufacturerDescriptorHeader.stringDescriptor, value, sizeof(wchar_t) * SIZEOF_ARRAY(stringManufacturerDescriptorHeader.stringDescriptor));
-        break;
-
-    case TinyCLR_UsbClient_StringDescriptorType::ProductName:
-        memcpy(&stringProductNameDescriptorHeader.stringDescriptor, value, sizeof(wchar_t) * SIZEOF_ARRAY(stringManufacturerDescriptorHeader.stringDescriptor));
-        break;
-
-    case TinyCLR_UsbClient_StringDescriptorType::DisplayName:
-        memcpy(&stringDisplayNameDescriptorHeader.stringDescriptor, value, sizeof(wchar_t) * SIZEOF_ARRAY(stringManufacturerDescriptorHeader.stringDescriptor));
-        break;
-
-    case TinyCLR_UsbClient_StringDescriptorType::FriendlyName:
-        memcpy(&stringFriendlyNameDescriptorHeader.stringDescriptor, value, sizeof(wchar_t) * SIZEOF_ARRAY(stringManufacturerDescriptorHeader.stringDescriptor));
-        break;
-    }
-
-    return TinyCLR_Result::Success;
-}
-
-void STM32F4_UsbClient_Reset() {
-    for (auto pipe = 0; pipe < STM32F4_USB_QUEUE_SIZE; pipe++) {
-        STM32F4_UsbClient_Close(&usbClientProvider, pipe);
-    }
-
-    STM32F4_UsbClient_Release(&usbClientProvider);
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-/// STM32F4 USB Hardware driver
-///////////////////////////////////////////////////////////////////////////////////////////
 #define OTG_FS_BASE           (0x50000000)
 #define OTG_FS                ((OTG_TypeDef *) OTG_FS_BASE)
 
@@ -1801,6 +141,12 @@ void STM32F4_UsbClient_Reset() {
 #define USB_OTG_NUM_FIFOS                8
 #define USB_OTG_NUM_CHANNELS            16
 
+#define USB_MAX_EP_SIZE            64 // maximum control channel packet size
+#define USB_MAX_EP0_SIZE            8 // default control channel packet size
+#define USB_FS_MAX_EP_COUNT         4 // OTG FS supports 4 endpoints
+
+
+
 typedef struct {
     // global registers
     __IO uint32_t GOTGCTL;
@@ -1897,6 +243,11 @@ typedef struct {
     uint8_t     previousDeviceState;
 
 } STM32F4_UsbClientController;
+
+static const STM32F4_Gpio_Pin g_STM32F4_Usb_Dm_Pins[] = STM32F4_USB_DM_PINS;
+static const STM32F4_Gpio_Pin g_STM32F4_Usb_Dp_Pins[] = STM32F4_USB_DP_PINS;
+static const STM32F4_Gpio_Pin g_STM32F4_Usb_Vb_Pins[] = STM32F4_USB_VB_PINS;
+static const STM32F4_Gpio_Pin g_STM32F4_Usb_Id_Pins[] = STM32F4_USB_ID_PINS;
 
 void STM32F4_UsbClient_ProtectPins(int32_t controller, bool On);
 void STM32F4_UsbClient_Interrupt(void* param);
@@ -2047,13 +398,13 @@ void STM32F4_UsbClient_ResetEvent(OTG_TypeDef* OTG, USB_CONTROLLER_STATE* usbSta
     OTG->DCFG &= ~OTG_DCFG_DAD; // reset device address
 
     /* clear all flags */
-    STM32F4_UsbClient_ClearEvent(usbState, 0xFFFFFFFF); // clear all events on all endpoints
+    UsbClient_ClearEvent(usbState, 0xFFFFFFFF); // clear all events on all endpoints
 
     usbState->firstGetDescriptor = true;
 
     usbState->deviceState = USB_DEVICE_STATE_DEFAULT;
     usbState->address = 0;
-    STM32F4_UsbClient_StateCallback(usbState);
+    UsbClient_StateCallback(usbState);
 }
 
 void STM32F4_UsbClient_EndpointRxInterrupt(OTG_TypeDef* OTG, USB_CONTROLLER_STATE* usbState, uint32_t ep, uint32_t count) {
@@ -2067,7 +418,7 @@ void STM32F4_UsbClient_EndpointRxInterrupt(OTG_TypeDef* OTG, USB_CONTROLLER_STAT
         usbState->dataSize = count;
     }
     else { // data endpoint
-        USB_PACKET64* Packet64 = STM32F4_UsbClient_RxEnqueue(usbState, ep, disableRx);
+        USB_PACKET64* Packet64 = UsbClient_RxEnqueue(usbState, ep, disableRx);
 
         if (disableRx) return;
 
@@ -2101,7 +452,7 @@ void STM32F4_UsbClient_EndpointInInterrupt(OTG_TypeDef* OTG, USB_CONTROLLER_STAT
         }
         else if (usbState->queues[ep] != 0 && usbState->isTxQueue[ep]) { // Tx data endpoint
 
-            USB_PACKET64* Packet64 = STM32F4_UsbClient_TxDequeue(usbState, ep);
+            USB_PACKET64* Packet64 = UsbClient_TxDequeue(usbState, ep);
 
             if (Packet64) {  // data to send
                 ps = (uint32_t*)Packet64->Buffer;
@@ -2129,7 +480,7 @@ void STM32F4_UsbClient_EndpointInInterrupt(OTG_TypeDef* OTG, USB_CONTROLLER_STAT
 
 void STM32F4_UsbClient_HandleSetup(OTG_TypeDef* OTG, USB_CONTROLLER_STATE* usbState) {
     /* send last setup packet to the upper layer */
-    uint8_t result = STM32F4_UsbClient_ControlCallback(usbState);
+    uint8_t result = UsbClient_ControlCallback(usbState);
 
     switch (result) {
 
@@ -2182,7 +533,7 @@ void STM32F4_UsbClient_EndpointOutInterrupt(OTG_TypeDef* OTG, USB_CONTROLLER_STA
         // Handle Setup data in upper layer
         STM32F4_UsbClient_HandleSetup(OTG, usbState);
     }
-    else if (usb_fifo_buffer_count[ep] < STM32F4_USB_FIFO_BUFFER_SIZE) {
+    else if (UsbClient_GetBufferCount(ep) < STM32F4_USB_FIFO_BUFFER_SIZE) {
         // enable endpoint
         OTG->DOEP[ep].TSIZ = OTG_DOEPTSIZ_PKTCNT_1 | usbState->maxPacketSize[ep];
         OTG->DOEP[ep].CTL |= OTG_DOEPCTL_EPENA | OTG_DOEPCTL_CNAK;
@@ -2255,7 +606,7 @@ void STM32F4_UsbClient_Interrupt(void* param) {
 
             usbState->deviceState = USB_DEVICE_STATE_SUSPENDED;
 
-            STM32F4_UsbClient_StateCallback(usbState);
+            UsbClient_StateCallback(usbState);
 
             OTG->GINTSTS = OTG_GINTSTS_USBSUSP; // clear interrupt
         }
@@ -2265,7 +616,7 @@ void STM32F4_UsbClient_Interrupt(void* param) {
 
             usbState->deviceState = usbClientController[controller].previousDeviceState;
 
-            STM32F4_UsbClient_StateCallback(usbState);
+            UsbClient_StateCallback(usbState);
 
             OTG->GINTSTS = OTG_GINTSTS_WKUPINT; // clear interrupt
         }
@@ -2286,7 +637,7 @@ bool STM32F4_UsbClient_StartOutput(USB_CONTROLLER_STATE* usbState, int32_t ep) {
 
     /* if the halt feature for this endpoint is set, then just clear all the characters */
     if (usbState->endpointStatus[ep] & USB_STATUS_ENDPOINT_HALT) {
-        STM32F4_UsbClient_ClearEndpoints(ep);
+        UsbClient_ClearEndpoints(ep);
 
         return true;
     }
@@ -2348,7 +699,7 @@ void STM32F4_UsbClient_ProtectPins(int32_t controller, bool on) {
         // clear USB Txbuffer
         for (int32_t ep = 1; ep < usbState->endpointCount; ep++) {
             if (usbState->queues[ep] && usbState->isTxQueue[ep]) {
-                STM32F4_UsbClient_ClearEndpoints(ep);
+                UsbClient_ClearEndpoints(ep);
             }
         }
 
@@ -2361,5 +712,5 @@ void STM32F4_UsbClient_ProtectPins(int32_t controller, bool on) {
 
     usbState->deviceState = on ? USB_DEVICE_STATE_ATTACHED : USB_DEVICE_STATE_DETACHED;
 
-    STM32F4_UsbClient_StateCallback(usbState);
+    UsbClient_StateCallback(usbState);
 }
